# Increase performance for the memory storage