Memory storage optimized:
```go
Benchmark_Memory/fiber_memory-12         	    7950	    153780 ns/op	     616 B/op	       0 allocs/op
Benchmark_Memory/fiber_memory-12         	    9714	    134795 ns/op	     657 B/op	       0 allocs/op
Benchmark_Memory/fiber_memory-12         	    9558	    122996 ns/op	     645 B/op	       0 allocs/op
Benchmark_Memory/fiber_memory-12         	    9392	    130914 ns/op	     635 B/op	       0 allocs/op
Benchmark_Memory/fiber_memory-12         	    8590	    123949 ns/op	     645 B/op	       0 allocs/op
Benchmark_Memory/fiber_memory-12         	    9662	    126010 ns/op	     703 B/op	       0 allocs/op
```

original storage:
```go
Benchmark_Memory/fiber_memory-12         	    3972	    282395 ns/op	     662 B/op	       0 allocs/op
Benchmark_Memory/fiber_memory-12         	    4204	    279402 ns/op	     723 B/op	       0 allocs/op
Benchmark_Memory/fiber_memory-12         	    4239	    278725 ns/op	     619 B/op	       0 allocs/op
Benchmark_Memory/fiber_memory-12         	    4239	    302900 ns/op	     717 B/op	       0 allocs/op
Benchmark_Memory/fiber_memory-12         	    4288	    282079 ns/op	     703 B/op	       0 allocs/op
Benchmark_Memory/fiber_memory-12         	    4221	    280041 ns/op	     721 B/op	       0 allocs/op
```